refactor to make terminal distinction clearer
