+++ src/ec2/activation.ts
@@ -4,16 +4,20 @@
  */
 import { ExtContext } from '../shared/extensions'
 import { Commands } from '../shared/vscode/commands2'
-import { tryConnect } from './commands'
 import { telemetry } from '../shared/telemetry/telemetry'
 import { Ec2InstanceNode } from './explorer/ec2InstanceNode'
+import { promptUserForEc2Selection } from './prompter'
+import { Ec2ConnectionManager } from './model'
 
 export async function activate(ctx: ExtContext): Promise<void> {
     ctx.extensionContext.subscriptions.push(
         Commands.register('aws.ec2.openTerminal', async (node?: Ec2InstanceNode) => {
             await telemetry.ec2_connectToInstance.run(async span => {
                 span.record({ ec2ConnectionType: 'ssm' })
-                await (node ? tryConnect(node.toSelection()) : tryConnect())
+                const selection = node ? node.toSelection() : await promptUserForEc2Selection()
+
+                const connectionManager = new Ec2ConnectionManager(selection.region)
+                await connectionManager.attemptToOpenEc2Terminal(selection)
             })
         }),
