+++ src/ec2/prompter.ts
@@ -7,6 +7,9 @@ import { RegionSubmenu, RegionSubmenuResponse } from '../shared/ui/common/region
 import { Ec2Selection, getInstancesFromRegion } from './utils'
 import { DataQuickPickItem } from '../shared/ui/pickerPrompter'
 import { Ec2Instance } from '../shared/clients/ec2Client'
+import { PromptResult } from '../shared/ui/prompter'
+import { isValidResponse } from '../shared/wizards/wizard'
+import { CancellationError } from '../shared/utilities/timeoutUtils'
 
 function asQuickpickItem(instance: Ec2Instance): DataQuickPickItem<string> {
     return {
@@ -14,6 +17,21 @@ function asQuickpickItem(instance: Ec2Instance): DataQuickPickItem<string> {
         detail: instance.InstanceId,
         data: instance.InstanceId,
     }
+}
+
+function getSelectionFromResponse(response: PromptResult<RegionSubmenuResponse<string>>): Ec2Selection {
+    if (isValidResponse(response)) {
+        return handleEc2ConnectPrompterResponse(response)
+    } else {
+        throw new CancellationError('user')
+    }
+}
+
+export async function promptUserForEc2Selection(): Promise<Ec2Selection> {
+    const prompter = createEc2ConnectPrompter()
+    const response = await prompter.prompt()
+
+    return getSelectionFromResponse(response)
 }
 
 export function handleEc2ConnectPrompterResponse(response: RegionSubmenuResponse<string>): Ec2Selection {
